references/requirements.txt
# Core
aiogram==3.15.0
uvloop==0.21.0
aiohttp==3.9.0

# Database
sqlalchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1

# Cache
redis==5.2.1

# Config
pydantic>=2.4.1,<2.10
pydantic-settings>=2.0.0,<3.0.0
python-dotenv==1.0.1

# Scheduler
apscheduler==3.10.4

# Logging
loguru==0.7.3

# I18n
babel==2.17.0

# Optional
sentry-sdk[loguru]==2.20.0
